The last of the renaming
Missed adding target to scsi_accel files to be explicit between
host and target scsi_accel files.
